Added go_to_page function and new export to configure the book’s initial state
Introduced a new export to configure the book’s initial state (open showing any page, or closed from the left or right).

New funcion added: "go_to_page". Interactive scenes can use it to navigate to the correct page (the page corresponds to the position in the array of pages starting at 1).